linux/requirements.txt
hidapi>=0.15.0